feat: add queue (doubly linked list implementation)

+++ include/dll.h
@@ -19,7 +19,7 @@ void dll_insert_at_head(dll_node_t **head, object_t *o);
 void dll_print(dll_node_t *head);
 dll_node_t *dll_node_at(dll_node_t *head, int index);
 int dll_search_node(dll_node_t *head, object_t *o);
-int dll_length(dll_node_t *head);
+size_t dll_length(dll_node_t *head);
 
 // void dll_insert_at(dll_node_t **head, int index, object_t *o);
 // void dll_delete_at(dll_node_t **head, int index);
